controller for cart actualizado

+++ controllers/cart/cart.js
@@ -7,9 +7,7 @@ const controller = {
         const { user } = req 
         try {
             let item = await Item.create({
-                quantity: 1,
                 user_id: user._id,
-                category_id: req.body.category_id,
                 car_id: req.body.car_id,
                 color_id: req.body.color_id,
                 rim_id: req.body.rim_id
@@ -29,7 +27,9 @@ const controller = {
         const { user } = req 
         try {
             let item = await Item.find( { user_id: user._id })
-                .populate("")
+                .populate("car_id  ")
+                .populate("color_id  ")
+                .populate("rim_id  ")
             return res
                 .status(200)
                 .json({
@@ -40,27 +40,27 @@ const controller = {
         }
     },
 
-    update: async (req,res,next) => {
-        try {
-            let product = await Item.findByIdAndUpdate(
-                req.params.id,
-                { quantity: req.body.quantity}
-                )
-            if (product){
-                return res 
-                    .status(200)
-                    .json({
-                        message: 'Quantity updated',
-                    })
-            }
-        } catch (error) {
-            next(error)
-        }
-    },
+    // update: async (req,res,next) => {
+    //     try {
+    //         let product = await Item.findByIdAndUpdate(
+    //             req.params.id,
+    //             { quantity: req.body.quantity}
+    //             )
+    //         if (product){
+    //             return res 
+    //                 .status(200)
+    //                 .json({
+    //                     message: 'Quantity updated',
+    //                 })
+    //         }
+    //     } catch (error) {
+    //         next(error)
+    //     }
+    // },
 
     deleteOne: async (req,res,next) => {
         try {
-            let item = await Item.deleteOne( {_id: req.params.id})
+            let item = await Item.deleteOne( {_id: req.params.id, user_id: req.user._id})
             if( item ){
                 return res
                     .status(200)
